# api-docs.yaml
openapi: 3.1.0
info:
  title: GBCC Website API
  description: |
    REST API интернет-магазина / каталога GBCC: аутентификация (JWT в HttpOnly cookies),
    аккаунты, каталог товаров и файлы, новости, акции (скидки), заказы, обращения в поддержку.

    **Авторизация в Swagger UI:** схема `accessTokenCookie` — cookie `GBCC_ACCESS_TOKEN`
    (после успешного `POST /auth/login` в том же браузере cookie подставляется автоматически;
    при тестах из Postman — передайте cookie вручную).

    **Гостевой токен поддержки:** схема `supportGuestToken` — заголовок `X-Support-Token`
    (значение из `guestAccessToken` после создания обращения без входа).
  version: 0.2.0
servers:
- url: http://localhost:8080
  description: Local Dev Server
...
paths:
...
  /news/{newsId}:
    get:
      tags:
      - News
      summary: Получить новость по id
      description: Публично. Не-менеджеры видят только опубликованные и с датой публикации
        не в будущем.
      operationId: getById_2
      parameters:
      - name: newsId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NewsResponseDto"
        "404":
          description: Нет доступа или не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
      - News
      summary: Полное обновление новости
      description: JWT. ADMIN/OWNER.
      operationId: update_2
      parameters:
      - name: newsId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNewsRequestDto"
        required: true
      responses:
        "200":
          description: Обновлено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NewsResponseDto"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет прав
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
    patch:
      tags:
      - News
      summary: Частичное обновление новости
      description: JWT. ADMIN/OWNER.
      operationId: patch_2
      parameters:
      - name: newsId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchNewsRequestDto"
        required: true
      responses:
        "200":
          description: Обновлено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NewsResponseDto"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет прав
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
...
  /news:
    get:
      tags:
      - News
      summary: Поиск новостей
      description: "Параметры: query (текст), isPublished, category (рубрика). Паг\
        инация Spring."
      operationId: search_3
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: isPublished
        in: query
        required: false
        schema:
          type: boolean
      - name: category
        in: query
        required: false
        schema:
          type: string
          enum:
          - PRODUCTS
          - PRODUCTION
          - SUPPLIES
          - PROJECTS
          - CLIENTS
          - PARTNERS
          - PROMOTIONS
      responses:
        "200":
          description: Страница новостей
          content:
            application/json: {}
    post:
      tags:
      - News
      summary: Создать новость
      description: Требует JWT. Только роли ADMIN или OWNER.
      operationId: create_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequestDto"
        required: true
      responses:
        "201":
          description: Создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NewsResponseDto"
        "400":
          description: Валидация
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
...
components:
  schemas:
...
    ApiErrorField:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ApiErrorResponse:
      type: object
      description: Единый формат ошибки API (см. GlobalExceptionHandler)
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorField"
...
    UpdateNewsRequestDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        previewText:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        category:
          type: string
          enum:
          - PRODUCTS
          - PRODUCTION
          - SUPPLIES
          - PROJECTS
          - CLIENTS
          - PARTNERS
          - PROMOTIONS
        coverFileId:
          type: string
          format: uuid
        isPublished:
          type: boolean
        publishedAt:
          type: string
          format: date-time
      required:
      - category
      - content
      - isPublished
      - previewText
      - title
    NewsResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        previewText:
          type: string
        content:
          type: string
        coverFileId:
          type: string
          format: uuid
        category:
          type: string
          enum:
          - PRODUCTS
          - PRODUCTION
          - SUPPLIES
          - PROJECTS
          - CLIENTS
          - PARTNERS
          - PROMOTIONS
        isPublished:
          type: boolean
        publishedAt:
          type: string
          format: date-time
        createdByAccountId:
          type: string
          format: uuid
        updatedByAccountId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateNewsRequestDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        previewText:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        category:
          type: string
          enum:
          - PRODUCTS
          - PRODUCTION
          - SUPPLIES
          - PROJECTS
          - CLIENTS
          - PARTNERS
          - PROMOTIONS
        coverFileId:
          type: string
          format: uuid
        isPublished:
          type: boolean
        publishedAt:
          type: string
          format: date-time
      required:
      - category
      - content
      - previewText
      - title
...
    PatchNewsRequestDto:
      type: object
      properties:
        title:
          type: string
        previewText:
          type: string
        content:
          type: string
        category:
          type: string
          enum:
          - PRODUCTS
          - PRODUCTION
          - SUPPLIES
          - PROJECTS
          - CLIENTS
          - PARTNERS
          - PROMOTIONS
        coverFileId:
          type: string
          format: uuid
        isPublished:
          type: boolean
        publishedAt:
          type: string
          format: date-time
...
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      description: |
        JWT access token (HttpOnly cookie). Выставляется после POST /auth/login.
        В Swagger UI: Authorize → значение токена (если копируете из DevTools) или вход через браузер на том же origin.
      name: GBCC_ACCESS_TOKEN
      in: cookie
    supportGuestToken:
      type: apiKey
      description: |
        Токен гостя для обращения в поддержку без аккаунта.
        Возвращается в теле ответа POST /support/conversations (поле guestAccessToken).
      name: X-Support-Token
      in: header